Move priv methods to Pantheon module
