+++ eea/src/main/java/org/web3j/protocol/eea/JsonRpc2_0Eea.java
@@ -12,20 +12,12 @@
  */
 package org.web3j.protocol.eea;
 
-import java.util.Arrays;
 import java.util.Collections;
-import java.util.List;
 
 import org.web3j.protocol.Web3jService;
-import org.web3j.protocol.admin.methods.response.BooleanResponse;
 import org.web3j.protocol.core.JsonRpc2_0Web3j;
 import org.web3j.protocol.core.Request;
-import org.web3j.protocol.core.methods.response.EthGetTransactionCount;
 import org.web3j.protocol.core.methods.response.EthSendTransaction;
-import org.web3j.protocol.eea.response.EeaCreatePrivacyGroup;
-import org.web3j.protocol.eea.response.EeaFindPrivacyGroup;
-import org.web3j.protocol.eea.response.EeaGetPrivacyPrecompileAddress;
-import org.web3j.protocol.eea.response.EeaGetPrivateTransaction;
 import org.web3j.protocol.eea.response.EeaGetTransactionReceipt;
 
 public class JsonRpc2_0Eea extends JsonRpc2_0Web3j implements Eea {
@@ -51,65 +43,5 @@ public class JsonRpc2_0Eea extends JsonRpc2_0Web3j implements Eea {
                 Collections.singletonList(transactionHash),
                 web3jService,
                 EeaGetTransactionReceipt.class);
-    }
-
-    @Override
-    public Request<?, EthGetTransactionCount> eeaGetTransactionCount(
-            final String address, final String privacyGroupId) {
-        return new Request<>(
-                "eea_getTransactionCount",
-                Arrays.asList(address, privacyGroupId),
-                web3jService,
-                EthGetTransactionCount.class);
-    }
-
-    @Override
-    public Request<?, EeaGetPrivateTransaction> eeaGetPrivateTransaction(final String enclaveKey) {
-        return new Request<>(
-                "eea_getPrivateTransaction",
-                Collections.singletonList(enclaveKey),
-                web3jService,
-                EeaGetPrivateTransaction.class);
-    }
-
-    @Override
-    public Request<?, EeaGetPrivacyPrecompileAddress> eeaGetPrivacyPrecompileAddress() {
-        return new Request<>(
-                "eea_getPrivacyPrecompileAddress",
-                Collections.emptyList(),
-                web3jService,
-                EeaGetPrivacyPrecompileAddress.class);
-    }
-
-    @Override
-    public Request<?, EeaCreatePrivacyGroup> eeaCreatePrivacyGroup(
-            final String from,
-            final String name,
-            final String description,
-            final List<String> addresses) {
-        return new Request<>(
-                "eea_createPrivacyGroup",
-                Arrays.asList(from, name, description, addresses),
-                web3jService,
-                EeaCreatePrivacyGroup.class);
-    }
-
-    @Override
-    public Request<?, EeaFindPrivacyGroup> eeaFindPrivacyGroup(final List<String> addresses) {
-        return new Request<>(
-                "eea_findPrivacyGroup",
-                Collections.singletonList(addresses),
-                web3jService,
-                EeaFindPrivacyGroup.class);
-    }
-
-    @Override
-    public Request<?, BooleanResponse> eeaDeletePrivacyGroup(
-            final String from, final String privacyGroupId) {
-        return new Request<>(
-                "eea_deletePrivacyGroup",
-                Arrays.asList(from, privacyGroupId),
-                web3jService,
-                BooleanResponse.class);
     }
 }
