+++ eea/src/test/java/org/web3j/protocol/eea/ResponseTest.java
@@ -12,25 +12,15 @@
  */
 package org.web3j.protocol.eea;
 
-import java.util.Arrays;
 import java.util.Collections;
-import java.util.Optional;
 
 import org.junit.Test;
 
 import org.web3j.protocol.ResponseTester;
-import org.web3j.protocol.admin.methods.response.BooleanResponse;
 import org.web3j.protocol.core.methods.response.EthSendTransaction;
 import org.web3j.protocol.core.methods.response.Log;
-import org.web3j.protocol.eea.response.EeaCreatePrivacyGroup;
-import org.web3j.protocol.eea.response.EeaFindPrivacyGroup;
-import org.web3j.protocol.eea.response.EeaGetPrivacyPrecompileAddress;
-import org.web3j.protocol.eea.response.EeaGetPrivateTransaction;
 import org.web3j.protocol.eea.response.EeaGetTransactionReceipt;
-import org.web3j.protocol.eea.response.PrivacyGroup;
-import org.web3j.protocol.eea.response.PrivateTransactionLegacy;
 import org.web3j.protocol.eea.response.PrivateTransactionReceipt;
-import org.web3j.protocol.eea.response.PrivateTransactionWithPrivacyGroup;
 
 import static org.hamcrest.CoreMatchers.equalTo;
 import static org.hamcrest.CoreMatchers.is;
@@ -106,208 +96,5 @@ public class ResponseTest extends ResponseTester {
         assertThat(
                 eeaGetTransactionReceipt.getTransactionReceipt().get(),
                 equalTo(transactionReceipt));
-    }
-
-    @Test
-    public void testEeaGetPrivateTransactionLegacy() {
-        // CHECKSTYLE:OFF
-        buildResponse(
-                "{\n"
-                        + "    \"id\":1,\n"
-                        + "    \"jsonrpc\":\"2.0\",\n"
-                        + "    \"result\": {\n"
-                        + "        \"hash\":\"0xc6ef2fc5426d6ad6fd9e2a26abeab0aa2411b7ab17f30a99d3cb96aed1d1055b\",\n"
-                        + "        \"nonce\":\"0x\",\n"
-                        + "        \"from\":\"0x407d73d8a49eeb85d32cf465507dd71d507100c1\",\n"
-                        + "        \"to\":\"0x85h43d8a49eeb85d32cf465507dd71d507100c1\",\n"
-                        + "        \"value\":\"0x7f110\",\n"
-                        + "        \"gas\": \"0x7f110\",\n"
-                        + "        \"gasPrice\":\"0x09184e72a000\",\n"
-                        + "        \"input\":\"0x603880600c6000396000f300603880600c6000396000f3603880600c6000396000f360\",\n"
-                        + "        \"r\":\"0xf115cc4d7516dd430046504e1c888198e0323e8ded016d755f89c226ba3481dc\",\n"
-                        + "        \"s\":\"0x4a2ae8ee49f1100b5c0202b37ed8bacf4caeddebde6b7f77e12e7a55893e9f62\",\n"
-                        + "        \"v\":0,\n"
-                        + "        \"privateFrom\":\"A1aVtMxLCUHmBVHXoZzzBgPbW/wj5axDpW9X8l91SGo=\",\n"
-                        + "        \"privateFor\":[\"A1aVtMxLCUHmBVHXoZzzBgPbW/wj5axDpW9X8l91SGo=\",\"Ko2bVqD+nNlNYL5EE7y3IdOnviftjiizpjRt+HTuFBs=\"],\n"
-                        + "        \"restriction\":\"restricted\""
-                        + "  }\n"
-                        + "}");
-        PrivateTransactionLegacy privateTransaction =
-                new PrivateTransactionLegacy(
-                        "0xc6ef2fc5426d6ad6fd9e2a26abeab0aa2411b7ab17f30a99d3cb96aed1d1055b",
-                        "0x",
-                        "0x407d73d8a49eeb85d32cf465507dd71d507100c1",
-                        "0x85h43d8a49eeb85d32cf465507dd71d507100c1",
-                        "0x7f110",
-                        "0x7f110",
-                        "0x09184e72a000",
-                        "0x603880600c6000396000f300603880600c6000396000f3603880600c6000396000f360",
-                        "0xf115cc4d7516dd430046504e1c888198e0323e8ded016d755f89c226ba3481dc",
-                        "0x4a2ae8ee49f1100b5c0202b37ed8bacf4caeddebde6b7f77e12e7a55893e9f62",
-                        (byte) 0,
-                        "A1aVtMxLCUHmBVHXoZzzBgPbW/wj5axDpW9X8l91SGo=",
-                        Arrays.asList(
-                                "A1aVtMxLCUHmBVHXoZzzBgPbW/wj5axDpW9X8l91SGo=",
-                                "Ko2bVqD+nNlNYL5EE7y3IdOnviftjiizpjRt+HTuFBs="),
-                        "restricted");
-        // CHECKSTYLE:ON
-
-        EeaGetPrivateTransaction eeaPrivateTransaction =
-                deserialiseResponse(EeaGetPrivateTransaction.class);
-        assertThat(
-                eeaPrivateTransaction.getPrivateTransaction().get(), equalTo(privateTransaction));
-    }
-
-    @Test
-    public void testEeaGetPrivateTransactionPrivacyGroup() {
-        // CHECKSTYLE:OFF
-        buildResponse(
-                "{\n"
-                        + "    \"id\":1,\n"
-                        + "    \"jsonrpc\":\"2.0\",\n"
-                        + "    \"result\": {\n"
-                        + "        \"hash\":\"0xc6ef2fc5426d6ad6fd9e2a26abeab0aa2411b7ab17f30a99d3cb96aed1d1055b\",\n"
-                        + "        \"nonce\":\"0x\",\n"
-                        + "        \"from\":\"0x407d73d8a49eeb85d32cf465507dd71d507100c1\",\n"
-                        + "        \"to\":\"0x85h43d8a49eeb85d32cf465507dd71d507100c1\",\n"
-                        + "        \"value\":\"0x7f110\",\n"
-                        + "        \"gas\": \"0x7f110\",\n"
-                        + "        \"gasPrice\":\"0x09184e72a000\",\n"
-                        + "        \"input\":\"0x603880600c6000396000f300603880600c6000396000f3603880600c6000396000f360\",\n"
-                        + "        \"r\":\"0xf115cc4d7516dd430046504e1c888198e0323e8ded016d755f89c226ba3481dc\",\n"
-                        + "        \"s\":\"0x4a2ae8ee49f1100b5c0202b37ed8bacf4caeddebde6b7f77e12e7a55893e9f62\",\n"
-                        + "        \"v\":0,\n"
-                        + "        \"privateFrom\":\"A1aVtMxLCUHmBVHXoZzzBgPbW/wj5axDpW9X8l91SGo=\",\n"
-                        + "        \"privacyGroupId\":\"A1aVtMxLCUHmBVHXoZzzBgPbW/wj5axDpW9X8l91SGo=\",\n"
-                        + "        \"restriction\":\"restricted\""
-                        + "  }\n"
-                        + "}");
-        PrivateTransactionWithPrivacyGroup privateTransaction =
-                new PrivateTransactionWithPrivacyGroup(
-                        "0xc6ef2fc5426d6ad6fd9e2a26abeab0aa2411b7ab17f30a99d3cb96aed1d1055b",
-                        "0x",
-                        "0x407d73d8a49eeb85d32cf465507dd71d507100c1",
-                        "0x85h43d8a49eeb85d32cf465507dd71d507100c1",
-                        "0x7f110",
-                        "0x7f110",
-                        "0x09184e72a000",
-                        "0x603880600c6000396000f300603880600c6000396000f3603880600c6000396000f360",
-                        "0xf115cc4d7516dd430046504e1c888198e0323e8ded016d755f89c226ba3481dc",
-                        "0x4a2ae8ee49f1100b5c0202b37ed8bacf4caeddebde6b7f77e12e7a55893e9f62",
-                        (byte) 0,
-                        "A1aVtMxLCUHmBVHXoZzzBgPbW/wj5axDpW9X8l91SGo=",
-                        "A1aVtMxLCUHmBVHXoZzzBgPbW/wj5axDpW9X8l91SGo=",
-                        "restricted");
-        // CHECKSTYLE:ON
-
-        EeaGetPrivateTransaction eeaPrivateTransaction =
-                deserialiseResponse(EeaGetPrivateTransaction.class);
-        assertThat(
-                eeaPrivateTransaction.getPrivateTransaction().get(), equalTo(privateTransaction));
-    }
-
-    @Test
-    public void testEeaGetPrivateTransactionNull() {
-        buildResponse("{\n" + "  \"result\": null\n" + "}");
-
-        EeaGetPrivateTransaction eeaPrivateTransaction =
-                deserialiseResponse(EeaGetPrivateTransaction.class);
-        assertThat(eeaPrivateTransaction.getPrivateTransaction(), is(Optional.empty()));
-    }
-
-    @Test
-    public void testEeaGetPrivacyPrecompileAddress() {
-        // CHECKSTYLE:OFF
-        buildResponse(
-                "{\n"
-                        + "    \"jsonrpc\": \"2.0\",\n"
-                        + "    \"id\": 1,\n"
-                        + "    \"result\": \"0xb60e8dd61c5d32be8058bb8eb970870f07233155\"\n"
-                        + "}");
-        // CHECKSTYLE:ON
-
-        EeaGetPrivacyPrecompileAddress eeaGetPrivacyPrecompileAddress =
-                deserialiseResponse(EeaGetPrivacyPrecompileAddress.class);
-        assertThat(
-                eeaGetPrivacyPrecompileAddress.getAddress(),
-                is("0xb60e8dd61c5d32be8058bb8eb970870f07233155"));
-    }
-
-    @Test
-    public void testEeaCreatePrivacyGroup() {
-        // CHECKSTYLE:OFF
-        buildResponse(
-                "{\n"
-                        + "    \"jsonrpc\": \"2.0\",\n"
-                        + "    \"id\": 1,\n"
-                        + "    \"result\": \"68/Cq0mVjB8FbXDLE1tbDRAvD/srluIok137uFOaClPU/dLFW34ovZebW+PTzy9wUawTXw==\"\n"
-                        + "}");
-        // CHECKSTYLE:ON
-
-        EeaCreatePrivacyGroup eeaCreatePrivacyGroup =
-                deserialiseResponse(EeaCreatePrivacyGroup.class);
-        assertThat(
-                eeaCreatePrivacyGroup.getPrivacyGroupId(),
-                is("68/Cq0mVjB8FbXDLE1tbDRAvD/srluIok137uFOaClPU/dLFW34ovZebW+PTzy9wUawTXw=="));
-    }
-
-    @Test
-    public void testEeaDeletePrivacyGroup() {
-        // CHECKSTYLE:OFF
-        buildResponse(
-                "{\n"
-                        + "    \"jsonrpc\": \"2.0\",\n"
-                        + "    \"id\": 1,\n"
-                        + "    \"result\": \"true\"\n"
-                        + "}");
-        // CHECKSTYLE:ON
-
-        BooleanResponse eeaDeletePrivacyGroup = deserialiseResponse(BooleanResponse.class);
-        assertThat(eeaDeletePrivacyGroup.success(), is(true));
-    }
-
-    @Test
-    public void testEeaFindPrivacyGroup() {
-        // CHECKSTYLE:OFF
-        buildResponse(
-                "{\n"
-                        + "    \"jsonrpc\": \"2.0\",\n"
-                        + "    \"id\": 1,\n"
-                        + "    \"result\": [\n"
-                        + "         {\n"
-                        + "            \"privacyGroupId\":\"A1aVtMxLCUHmBVHXoZzzBgPbW/wj5axDpW9X8l91SGo=\",\n"
-                        + "            \"name\":\"PrivacyGroupName\",\n"
-                        + "            \"description\":\"PrivacyGroupDescription\",\n"
-                        + "            \"type\":\"LEGACY\",\n"
-                        + "            \"members\": [\"A1aVtMxLCUHmBVHXoZzzBgPbW/wj5axDpW9X8l91SGo=\"]\n"
-                        + "         },\n"
-                        + "         {\n"
-                        + "            \"privacyGroupId\":\"A1aVtMxLCUHmBVHXoZzzBgPbW/wj5axDpW9X8l91SGo=\",\n"
-                        + "            \"name\":\"PrivacyGroupName\",\n"
-                        + "            \"description\":\"PrivacyGroupDescription\",\n"
-                        + "            \"type\":\"PANTHEON\",\n"
-                        + "            \"members\": [\"A1aVtMxLCUHmBVHXoZzzBgPbW/wj5axDpW9X8l91SGo=\"]\n"
-                        + "         }\n"
-                        + "    ]\n"
-                        + "}");
-        // CHECKSTYLE:ON
-        PrivacyGroup privacyGroup1 =
-                new PrivacyGroup(
-                        "A1aVtMxLCUHmBVHXoZzzBgPbW/wj5axDpW9X8l91SGo=",
-                        PrivacyGroup.Type.LEGACY,
-                        "PrivacyGroupName",
-                        "PrivacyGroupDescription",
-                        Collections.singletonList("A1aVtMxLCUHmBVHXoZzzBgPbW/wj5axDpW9X8l91SGo="));
-        PrivacyGroup privacyGroup2 =
-                new PrivacyGroup(
-                        "A1aVtMxLCUHmBVHXoZzzBgPbW/wj5axDpW9X8l91SGo=",
-                        PrivacyGroup.Type.PANTHEON,
-                        "PrivacyGroupName",
-                        "PrivacyGroupDescription",
-                        Collections.singletonList("A1aVtMxLCUHmBVHXoZzzBgPbW/wj5axDpW9X8l91SGo="));
-
-        EeaFindPrivacyGroup eeaFindPrivacyGroup = deserialiseResponse(EeaFindPrivacyGroup.class);
-        assertThat(
-                eeaFindPrivacyGroup.getGroups(), is(Arrays.asList(privacyGroup1, privacyGroup2)));
     }
 }
