+++ pantheon/src/main/java/org/web3j/protocol/pantheon/response/PrivCreatePrivacyGroup.java
@@ -10,11 +10,11 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package org.web3j.protocol.eea.response;
+package org.web3j.protocol.pantheon.response;
 
 import org.web3j.protocol.core.Response;
 
-public class EeaCreatePrivacyGroup extends Response<String> {
+public class PrivCreatePrivacyGroup extends Response<String> {
     public String getPrivacyGroupId() {
         return getResult();
     }
