+++ pantheon/src/main/java/org/web3j/protocol/pantheon/response/PrivGetPrivacyPrecompileAddress.java
@@ -10,14 +10,12 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package org.web3j.protocol.eea.response;
-
-import java.util.List;
+package org.web3j.protocol.pantheon.response;
 
 import org.web3j.protocol.core.Response;
 
-public class EeaFindPrivacyGroup extends Response<List<PrivacyGroup>> {
-    public List<PrivacyGroup> getGroups() {
+public class PrivGetPrivacyPrecompileAddress extends Response<String> {
+    public String getAddress() {
         return getResult();
     }
 }
