+++ pantheon/src/main/java/org/web3j/protocol/pantheon/response/PrivGetPrivateTransaction.java
@@ -10,13 +10,14 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package org.web3j.protocol.eea.response;
+package org.web3j.protocol.pantheon.response;
 
 import java.util.Optional;
 
 import org.web3j.protocol.core.Response;
+import org.web3j.protocol.eea.response.PrivateTransaction;
 
-public class EeaGetPrivateTransaction extends Response<PrivateTransaction> {
+public class PrivGetPrivateTransaction extends Response<PrivateTransaction> {
     public Optional<PrivateTransaction> getPrivateTransaction() {
         return Optional.ofNullable(getResult());
     }
