+++ pantheon/src/main/java/org/web3j/tx/response/PrivateTransactionReceiptProcessor.java
@@ -15,24 +15,24 @@ package org.web3j.tx.response;
 import java.io.IOException;
 import java.util.Optional;
 
-import org.web3j.protocol.eea.Eea;
 import org.web3j.protocol.eea.response.EeaGetTransactionReceipt;
 import org.web3j.protocol.eea.response.PrivateTransactionReceipt;
 import org.web3j.protocol.exceptions.TransactionException;
+import org.web3j.protocol.pantheon.Pantheon;
 
 public abstract class PrivateTransactionReceiptProcessor extends TransactionReceiptProcessor {
-    private Eea eea;
+    private Pantheon pantheon;
 
-    public PrivateTransactionReceiptProcessor(Eea eea) {
-        super(eea);
-        this.eea = eea;
+    public PrivateTransactionReceiptProcessor(Pantheon pantheon) {
+        super(pantheon);
+        this.pantheon = pantheon;
     }
 
     @Override
     Optional<PrivateTransactionReceipt> sendTransactionReceiptRequest(String transactionHash)
             throws IOException, TransactionException {
         EeaGetTransactionReceipt transactionReceipt =
-                eea.eeaGetTransactionReceipt(transactionHash).send();
+                pantheon.eeaGetTransactionReceipt(transactionHash).send();
         if (transactionReceipt.hasError()) {
             throw new TransactionException(
                     "Error processing request: " + transactionReceipt.getError().getMessage());
