+++ pantheon/src/test/java/org/web3j/protocol/pantheon/RequestTest.java
@@ -12,6 +12,7 @@
  */
 package org.web3j.protocol.pantheon;
 
+import java.util.Arrays;
 import java.util.HashMap;
 import java.util.Map;
 
@@ -113,5 +114,87 @@ public class RequestTest extends RequestTester {
                         + "\"disableStorage\":false,"
                         + "\"disableStack\":false}],"
                         + "\"id\":1}");
+    }
+
+    @Test
+    public void testEeaGetTransactionCount() throws Exception {
+        web3j.privGetTransactionCount("0x407d73d8a49eeb85d32cf465507dd71d507100c1", "0x0").send();
+
+        verifyResult(
+                "{\"jsonrpc\":\"2.0\",\"method\":\"eea_getTransactionCount\","
+                        + "\"params\":[\"0x407d73d8a49eeb85d32cf465507dd71d507100c1\",\"0x0\"],"
+                        + "\"id\":1}");
+    }
+
+    @Test
+    public void testEeaGetPrivateTransaction() throws Exception {
+        web3j.privGetPrivateTransaction("EnclaveKey").send();
+
+        // CHECKSTYLE:OFF
+        verifyResult(
+                "{\"jsonrpc\":\"2.0\",\"method\":\"eea_getPrivateTransaction\","
+                        + "\"params\":[\"EnclaveKey\"],"
+                        + "\"id\":1}");
+        // CHECKSTYLE:ON
+    }
+
+    @Test
+    public void testEeaGetPrivacyPrecompileAddress() throws Exception {
+        web3j.privGetPrivacyPrecompileAddress().send();
+
+        // CHECKSTYLE:OFF
+        verifyResult(
+                "{\"jsonrpc\":\"2.0\",\"method\":\"eea_getPrivacyPrecompileAddress\","
+                        + "\"params\":[],"
+                        + "\"id\":1}");
+        // CHECKSTYLE:ON
+    }
+
+    @Test
+    public void testEeaCreatePrivacyGroup() throws Exception {
+        web3j.privCreatePrivacyGroup(
+                        Arrays.asList(
+                                "A1aVtMxLCUHmBVHXoZzzBgPbW/wj5axDpW9X8l91SGo=",
+                                "Ko2bVqD+nNlNYL5EE7y3IdOnviftjiizpjRt+HTuFBs="),
+                        "testName",
+                        "testDesc")
+                .send();
+
+        // CHECKSTYLE:OFF
+        verifyResult(
+                "{\"jsonrpc\":\"2.0\",\"method\":\"eea_createPrivacyGroup\","
+                        + "\"params\":[\"A1aVtMxLCUHmBVHXoZzzBgPbW/wj5axDpW9X8l91SGo=\",\"testName\",\"testDesc\",[\"A1aVtMxLCUHmBVHXoZzzBgPbW/wj5axDpW9X8l91SGo=\",\"Ko2bVqD+nNlNYL5EE7y3IdOnviftjiizpjRt+HTuFBs=\"]],"
+                        + "\"id\":1}");
+        // CHECKSTYLE:ON
+    }
+
+    @Test
+    public void testEeaFindPrivacyGroup() throws Exception {
+        web3j.privFindPrivacyGroup(
+                        Arrays.asList(
+                                "A1aVtMxLCUHmBVHXoZzzBgPbW/wj5axDpW9X8l91SGo=",
+                                "A1aVtMxLCUHmBVHXoZzzBgPbW/wj5axDpW9X8l91SGo="))
+                .send();
+
+        // CHECKSTYLE:OFF
+        verifyResult(
+                "{\"jsonrpc\":\"2.0\",\"method\":\"eea_findPrivacyGroup\","
+                        + "\"params\":[[\"A1aVtMxLCUHmBVHXoZzzBgPbW/wj5axDpW9X8l91SGo=\",\"A1aVtMxLCUHmBVHXoZzzBgPbW/wj5axDpW9X8l91SGo=\"]],"
+                        + "\"id\":1}");
+        // CHECKSTYLE:ON
+    }
+
+    @Test
+    public void testEeaDeletePrivacyGroup() throws Exception {
+        web3j.privDeletePrivacyGroup(
+                        "68/Cq0mVjB8FbXDLE1tbDRAvD/srluIok137uFOaClPU/dLFW34ovZebW+PTzy9wUawTXw==")
+                .send();
+
+        // CHECKSTYLE:OFF
+        verifyResult(
+                "{\"jsonrpc\":\"2.0\",\"method\":\"eea_deletePrivacyGroup\","
+                        + "\"params\":[\"68/Cq0mVjB8FbXDLE1tbDRAvD/srluIok137uFOaClPU/dLFW34ovZebW+PTzy9wUawTXw==\"],"
+                        + "\"id\":1}");
+        // CHECKSTYLE:ON
     }
 }
